requirements.txt
torch
black
pre-commit
pytest
numpy
tqdm
matplotlib
pyinterpx
pyyaml
wandb